requirements.txt
requests
ollama
openai
pandas
pyyaml
tabulate
python-docx
pdfplumber
playwright
python-dotenv
pytest
anthropic
